@@ -22,22 +22,19 @@
   "homepage": "./",
   "private": true,
   "dependencies": {
-    "@tailwindcss/vite": "^4.1.7",
-    "@testing-library/jest-dom": "^6.6.3",
-    "@testing-library/react": "^16.3.0",
-    "@testing-library/user-event": "^14.6.1",
+    "@tailwindcss/vite": "^4.1.8",
     "@theme-toggles/react": "^4.1.0",
-    "@vitejs/plugin-react": "^4.4.1",
+    "@vitejs/plugin-react": "^4.5.1",
     "axios": "^1.9.0",
     "chart.js": "^4.4.9",
     "electron-is-dev": "^3.0.1",
     "electron-log": "^5.4.0",
-    "jssoup": "0.0.15",
     "lodash": "^4.17.21",
+    "node-html-parser": "^7.0.1",
     "react": "^19.1.0",
     "react-chartjs-2": "^5.3.0",
     "react-dom": "^19.1.0",
-    "tailwindcss": "^4.1.7",
+    "tailwindcss": "^4.1.8",
     "vite": "^6.3.5"
   },
   "scripts": {
@@ -64,9 +61,9 @@
     ]
   },
   "devDependencies": {
-    "electron": "^36.2.1",
+    "electron": "^36.3.2",
     "electron-builder": "^26.0.12",
-    "eslint": "^9.27.0",
+    "eslint": "^9.28.0",
     "eslint-plugin-react": "^7.37.5",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
